change package setup to include templatetags
