+++ setup.cfg
@@ -29,18 +29,15 @@ classifiers =
     Programming Language :: Python :: 3.9
 license_file = LICENSE
 
+[bdist_wheel]
+universal = 1
+
 [options]
-package_dir=
-    =src
-py_modules = django_htmx
-include_package_data = True
 install_requires =
     Django>=2.2
 python_requires = >=3.5
-zip_safe = False
+include_package_data = True
 
-[options.packages.find]
-where = src
 
 [flake8]
 max-line-length = 80
