+++ setup.py
@@ -1,3 +1,5 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
-setup()
+setup(
+    packages=find_packages()
+)
